+++ actionview/lib/action_view/helpers/form_options_helper.rb
@@ -794,7 +794,7 @@ module ActionView
         def extract_values_from_collection(collection, value_method, selected)
           if selected.is_a?(Proc)
             collection.map do |element|
-              element.public_send(value_method) if selected.call(element)
+              public_or_deprecated_send(element, value_method) if selected.call(element)
             end.compact
           else
             selected
@@ -802,7 +802,17 @@ module ActionView
         end
 
         def value_for_collection(item, value)
-          value.respond_to?(:call) ? value.call(item) : item.public_send(value)
+          value.respond_to?(:call) ? value.call(item) : public_or_deprecated_send(item, value)
+        end
+
+        def public_or_deprecated_send(item, value)
+          begin
+            item.public_send(value)
+          rescue NoMethodError
+            raise unless item.respond_to?(value, true) && !item.respond_to?(value)
+            ActiveSupport::Deprecation.warn "Using private methods in view's helpers is deprecated"
+            item.send(value)
+          end
         end
 
         def prompt_text(prompt)